ask db
